refactor(crypto): move generatePrivateKey into schnorr.ts
