+++ packages/zilliqa-js-crypto/src/schnorr.ts
@@ -7,7 +7,8 @@ import { BN } from '@zilliqa-js/util';
 import { randomBytes } from './random';
 import { Signature } from './signature';
 
-const curve = elliptic.ec('secp256k1').curve;
+const secp256k1 = elliptic.ec('secp256k1');
+const curve = secp256k1.curve;
 // Public key is a point (x, y) on the curve.
 // Each coordinate requires 32 bytes.
 // In its compressed form it suffices to store the x co-ordinate
@@ -21,6 +22,22 @@ const ALG_LEN = 16;
 // The length in bytes of entropy inputs to HMAC-DRBG
 const ENT_LEN = 32;
 
+const HEX_ENC: 'hex' = 'hex';
+
+/**
+ * generatePrivateKey
+ *
+ * @returns {string} - the hex-encoded private key
+ */
+export const generatePrivateKey = (): string => {
+  return secp256k1
+    .genKeyPair({
+      entropy: randomBytes(secp256k1.curve.n.byteLength()),
+      entropyEnc: HEX_ENC,
+      pers: 'zilliqajs+secp256k1+SHA256',
+    })
+    .getPrivate(HEX_ENC);
+};
 /**
  * Hash (r | M).
  * @param {Buffer} msg
